Ported delete tests
