+++ enginetest/engine_test.go
@@ -1015,143 +1015,6 @@ func TestDescribeNoPruneColumns(t *testing.T) {
 	require.Len(p.Schema(), 3)
 }
 
-func TestDeleteFrom(t *testing.T) {
-	var deletions = []struct {
-		deleteQuery    string
-		expectedDelete []sql.Row
-		selectQuery    string
-		expectedSelect []sql.Row
-	}{
-		{
-			"DELETE FROM mytable;",
-			[]sql.Row{{sql.NewOkResult(3)}},
-			"SELECT * FROM mytable;",
-			nil,
-		},
-		{
-			"DELETE FROM mytable WHERE i = 2;",
-			[]sql.Row{{sql.NewOkResult(1)}},
-			"SELECT * FROM mytable;",
-			[]sql.Row{{int64(1), "first row"}, {int64(3), "third row"}},
-		},
-		{
-			"DELETE FROM mytable WHERE i < 3;",
-			[]sql.Row{{sql.NewOkResult(2)}},
-			"SELECT * FROM mytable;",
-			[]sql.Row{{int64(3), "third row"}},
-		},
-		{
-			"DELETE FROM mytable WHERE i > 1;",
-			[]sql.Row{{sql.NewOkResult(2)}},
-			"SELECT * FROM mytable;",
-			[]sql.Row{{int64(1), "first row"}},
-		},
-		{
-			"DELETE FROM mytable WHERE i <= 2;",
-			[]sql.Row{{sql.NewOkResult(2)}},
-			"SELECT * FROM mytable;",
-			[]sql.Row{{int64(3), "third row"}},
-		},
-		{
-			"DELETE FROM mytable WHERE i >= 2;",
-			[]sql.Row{{sql.NewOkResult(2)}},
-			"SELECT * FROM mytable;",
-			[]sql.Row{{int64(1), "first row"}},
-		},
-		{
-			"DELETE FROM mytable WHERE s = 'first row';",
-			[]sql.Row{{sql.NewOkResult(1)}},
-			"SELECT * FROM mytable;",
-			[]sql.Row{{int64(2), "second row"}, {int64(3), "third row"}},
-		},
-		{
-			"DELETE FROM mytable WHERE s <> 'dne';",
-			[]sql.Row{{sql.NewOkResult(3)}},
-			"SELECT * FROM mytable;",
-			nil,
-		},
-		{
-			"DELETE FROM mytable WHERE s LIKE '%row';",
-			[]sql.Row{{sql.NewOkResult(3)}},
-			"SELECT * FROM mytable;",
-			nil,
-		},
-		{
-			"DELETE FROM mytable WHERE s = 'dne';",
-			[]sql.Row{{sql.NewOkResult(0)}},
-			"SELECT * FROM mytable;",
-			[]sql.Row{{int64(1), "first row"}, {int64(2), "second row"}, {int64(3), "third row"}},
-		},
-		{
-			"DELETE FROM mytable WHERE i = 'invalid';",
-			[]sql.Row{{sql.NewOkResult(0)}},
-			"SELECT * FROM mytable;",
-			[]sql.Row{{int64(1), "first row"}, {int64(2), "second row"}, {int64(3), "third row"}},
-		},
-		{
-			"DELETE FROM mytable ORDER BY i ASC LIMIT 2;",
-			[]sql.Row{{sql.NewOkResult(2)}},
-			"SELECT * FROM mytable;",
-			[]sql.Row{{int64(3), "third row"}},
-		},
-		{
-			"DELETE FROM mytable ORDER BY i DESC LIMIT 1;",
-			[]sql.Row{{sql.NewOkResult(1)}},
-			"SELECT * FROM mytable;",
-			[]sql.Row{{int64(1), "first row"}, {int64(2), "second row"}},
-		},
-		{
-			"DELETE FROM mytable ORDER BY i DESC LIMIT 1 OFFSET 1;",
-			[]sql.Row{{sql.NewOkResult(1)}},
-			"SELECT * FROM mytable;",
-			[]sql.Row{{int64(1), "first row"}, {int64(3), "third row"}},
-		},
-	}
-
-	for _, deletion := range deletions {
-		e, idxReg := NewEngine(t)
-		ctx := enginetest.NewCtx(idxReg)
-		enginetest.TestQuery(t, ctx, e, deletion.deleteQuery, deletion.expectedDelete)
-		enginetest.TestQuery(t, ctx, e, deletion.selectQuery, deletion.expectedSelect)
-	}
-}
-
-func TestDeleteFromErrors(t *testing.T) {
-	var expectedFailures = []struct {
-		name  string
-		query string
-	}{
-		{
-			"invalid table",
-			"DELETE FROM invalidtable WHERE x < 1;",
-		},
-		{
-			"invalid column",
-			"DELETE FROM mytable WHERE z = 'dne';",
-		},
-		{
-			"negative limit",
-			"DELETE FROM mytable LIMIT -1;",
-		},
-		{
-			"negative offset",
-			"DELETE FROM mytable LIMIT 1 OFFSET -1;",
-		},
-		{
-			"missing keyword from",
-			"DELETE mytable WHERE id = 1;",
-		},
-	}
-
-	for _, expectedFailure := range expectedFailures {
-		t.Run(expectedFailure.name, func(t *testing.T) {
-			e, idxReg := NewEngine(t)
-			_, _, err := e.Query(enginetest.NewCtx(idxReg), expectedFailure.query)
-			require.Error(t, err)
-		})
-	}
-}
-
 type mockSpan struct {
 	opentracing.Span
 	finished bool
